change code for the preferences-export
